Better commands to observe dictionary state

+++ dict.h
@@ -72,7 +72,7 @@ struct variable_t
 	char * name;
 	variable_t * next;
 	codeword_t cw[2]; // 0:pushVar, 1:assignVar
-	int value;
+	intptr_t value;
 };
 
 struct subdict
